Move Skeleton class to skeleton.py
